refactor(state): wire VerifyTrie through the backend-agnostic world-state module
Move the RunVerifyTrie diagnostic step from the trie-specific PruningTrieStoreModule into the general WorldStateModule, and split admin_verifyTrie out of IPruningTrieStateAdminRpcModule into its own backend-agnostic IVerifyTrieAdminRpcModule registered once for both backends. This makes admin_verifyTrie functional on flat nodes (it previously returned "disabled"); only admin_prune stays on the per-backend-swapped pruning admin module.
