+++ src/Nethermind/Nethermind.Db.Test/StandardDbInitializerTests.cs
@@ -86,7 +86,7 @@ public class StandardDbInitializerTests
             {
                 DownloadReceiptsInFastSync = useReceipts
             }))
-            .AddModule(new PruningTrieStoreModule(initConfig)) // For the full pruning db
+            .AddModule(new PruningTrieStoreModule()) // For the full pruning db
             .AddSingleton<IPruningConfig>(new PruningConfig())
             .AddSingleton<IDbConfig>(new DbConfig())
             .AddSingleton<IInitConfig>(initConfig)
