+++ src/Nethermind/Nethermind.Init/Modules/FlatWorldStateModule.cs
@@ -5,7 +5,6 @@ using System;
 using Autofac;
 using Nethermind.Api.Steps;
 using Nethermind.Blockchain;
-using Nethermind.Blockchain.Find;
 using Nethermind.Blockchain.FullPruning;
 using Nethermind.Blockchain.Synchronization;
 using Nethermind.Config;
@@ -111,7 +110,5 @@ public class FlatWorldStateModule(IFlatDbConfig flatDbConfig) : Module
     internal class PruningTrieStateAdminRpcModuleStub : IPruningTrieStateAdminRpcModule
     {
         public ResultWrapper<PruningStatus> admin_prune() => ResultWrapper<PruningStatus>.Success(PruningStatus.Disabled);
-
-        public ResultWrapper<string> admin_verifyTrie(BlockParameter block) => ResultWrapper<string>.Success("disabled");
     }
 }
