+++ src/Nethermind/Nethermind.Init/Modules/NethermindModule.cs
@@ -50,8 +50,8 @@ public class NethermindModule(ChainSpec chainSpec, IConfigProvider configProvide
                 configProvider.GetConfig<ISyncConfig>()
             ))
             .AddModule(new DbMonitoringModule())
-            .AddModule(new WorldStateModule())
-            .AddModule(new PruningTrieStoreModule(configProvider.GetConfig<IInitConfig>()))
+            .AddModule(new WorldStateModule(configProvider.GetConfig<IInitConfig>()))
+            .AddModule(new PruningTrieStoreModule())
             .AddModule(new FlatWorldStateModule(configProvider.GetConfig<IFlatDbConfig>()))
             .AddModule(new WorldStateDbDeciderModule())
             .AddModule(new PrewarmerModule(configProvider.GetConfig<IBlocksConfig>()))
