+++ src/Nethermind/Nethermind.Init/Modules/WorldStateModule.cs
@@ -3,6 +3,8 @@
 
 using System;
 using Autofac;
+using Nethermind.Api;
+using Nethermind.Api.Steps;
 using Nethermind.Blockchain;
 using Nethermind.Core;
 using Nethermind.JsonRpc.Modules;
@@ -12,9 +14,10 @@ using Nethermind.Trie.Pruning;
 
 namespace Nethermind.Init.Modules;
 
-public class WorldStateModule : Module
+public class WorldStateModule(IInitConfig initConfig) : Module
 {
-    protected override void Load(ContainerBuilder builder) =>
+    protected override void Load(ContainerBuilder builder)
+    {
         builder
             // Stub: overridden by WorldStateDbDeciderModule which selects patricia or flat at runtime.
             .AddSingleton<IWorldStateManager>(_ => throw new InvalidOperationException(
@@ -33,5 +36,13 @@ public class WorldStateModule : Module
 
             // Admin RPC surface is common to all backends; each backend registers its implementation.
             .RegisterSingletonJsonRpcModule<IPruningTrieStateAdminRpcModule>()
+
+            // Verify-trie admin RPC is backend-agnostic; a single implementation serves both backends.
+            .RegisterSingletonJsonRpcModule<IVerifyTrieAdminRpcModule, VerifyTrieAdminRpcModule>()
         ;
+
+        // Backend-agnostic diagnostic step; VerifyTrie resolves to whichever backend is active.
+        if (initConfig.DiagnosticMode == DiagnosticMode.VerifyTrie)
+            builder.AddStep(typeof(RunVerifyTrie));
+    }
 }
